+++ Vault/src/utopia/vault/model/immutable/access/NonDeprecatedAccess.scala
@@ -8,7 +8,7 @@ import utopia.vault.nosql.template.Deprecatable
  * @author Mikko Hilpinen
  * @since 11.1.2020, v1.4
  */
-@deprecated("Replaced with utopia.vault.nosql.access.template.NonDeprecatedAccess", "v1.4")
+@deprecated("Replaced with utopia.vault.nosql.access.template.model.NonDeprecatedAccess", "v1.4")
 trait NonDeprecatedAccess[+A] extends ConditionalAccess[A]
 {
 	// ABSTRACT	---------------------