+++ Vault/src/utopia/vault/nosql/access/single/SingleAccess.scala
@@ -2,7 +2,7 @@ package utopia.vault.nosql.access.single
 
 import utopia.vault.database.Connection
 import utopia.vault.model.immutable.Column
-import utopia.vault.nosql.access.template.FilterableAccess
+import utopia.vault.nosql.access.template.Access
 import utopia.vault.sql.OrderDirection.{Ascending, Descending}
 import utopia.vault.sql.{Condition, OrderBy, OrderDirection}
 
@@ -11,7 +11,7 @@ import utopia.vault.sql.{Condition, OrderBy, OrderDirection}
   * @author Mikko Hilpinen
   * @since 30.1.2020, v1.4
   */
-trait SingleAccess[+A, +Repr] extends FilterableAccess[Option[A], Repr]
+trait SingleAccess[+A] extends Access[Option[A]]
 {
 	// OTHER	---------------------
 	
@@ -23,14 +23,12 @@ trait SingleAccess[+A, +Repr] extends FilterableAccess[Option[A], Repr]
 	  */
 	def first(ordering: OrderBy, additionalCondition: Option[Condition])(implicit connection: Connection) =
 		read(mergeCondition(additionalCondition), Some(ordering))
-	
 	/**
 	  * @param ordering   Ordering used
 	  * @param connection Implicit database connection
 	  * @return The first item in specified ordering
 	  */
 	def first(ordering: OrderBy)(implicit connection: Connection): Option[A] = first(ordering, None)
-	
 	/**
 	  * @param ordering   Ordering used
 	  * @param condition  An additional search confition used
@@ -50,7 +48,6 @@ trait SingleAccess[+A, +Repr] extends FilterableAccess[Option[A], Repr]
 	  */
 	def top(orderColumn: Column, orderDirection: OrderDirection, additionalCondition: Option[Condition])
 	       (implicit connection: Connection) = first(OrderBy(orderColumn, orderDirection), additionalCondition)
-	
 	/**
 	  * The "top" value based on specified ordering
 	  * @param orderColumn    Ordering column
@@ -70,7 +67,6 @@ trait SingleAccess[+A, +Repr] extends FilterableAccess[Option[A], Repr]
 	  */
 	def minBy(column: Column, additionalCondition: Option[Condition])(implicit connection: Connection) =
 		top(column, Ascending, additionalCondition)
-	
 	/**
 	  * The minimum value based on specified ordering
 	  * @param column     Ordering column
@@ -78,7 +74,6 @@ trait SingleAccess[+A, +Repr] extends FilterableAccess[Option[A], Repr]
 	  * @return The smallest item based on provided ordering.
 	  */
 	def minBy(column: Column)(implicit connection: Connection): Option[A] = minBy(column, None)
-	
 	/**
 	  * The minimum value based on specified ordering
 	  * @param propertyName        Name of ordering property
@@ -88,7 +83,6 @@ trait SingleAccess[+A, +Repr] extends FilterableAccess[Option[A], Repr]
 	  */
 	def minBy(propertyName: String, additionalCondition: Option[Condition])(implicit connection: Connection): Option[A] =
 		minBy(table(propertyName), additionalCondition)
-	
 	/**
 	  * The minimum value based on specified ordering
 	  * @param propertyName Name of ordering property
@@ -106,7 +100,6 @@ trait SingleAccess[+A, +Repr] extends FilterableAccess[Option[A], Repr]
 	  */
 	def maxBy(column: Column, additionalCondition: Option[Condition])(implicit connection: Connection) =
 		top(column, Descending, additionalCondition)
-	
 	/**
 	  * The maximum value based on specified ordering
 	  * @param column     Ordering column
@@ -114,7 +107,6 @@ trait SingleAccess[+A, +Repr] extends FilterableAccess[Option[A], Repr]
 	  * @return The largest item based on provided ordering.
 	  */
 	def maxBy(column: Column)(implicit connection: Connection): Option[A] = maxBy(column, None)
-	
 	/**
 	  * The maximum value based on specified ordering
 	  * @param propertyName        Name of ordering property
@@ -124,7 +116,6 @@ trait SingleAccess[+A, +Repr] extends FilterableAccess[Option[A], Repr]
 	  */
 	def maxBy(propertyName: String, additionalCondition: Option[Condition])(implicit connection: Connection): Option[A] =
 		maxBy(table(propertyName), additionalCondition)
-	
 	/**
 	  * The maximum value based on specified ordering
 	  * @param propertyName Name of ordering property