+++ Vault/src/utopia/vault/nosql/access/single/model/SingleRowModelAccess.scala
@@ -1,11 +1,10 @@
 package utopia.vault.nosql.access.single.model
 
-import utopia.flow.datastructure.immutable.Value
-import utopia.vault.nosql.access.template.model.RowModelAccess
+import utopia.vault.nosql.view.RowFactoryView
 
 /**
   * Used for accessing individual models where each model occupies exactly one row
   * @author Mikko Hilpinen
   * @since 16.5.2020, v1.6
   */
-trait SingleRowModelAccess[+A] extends RowModelAccess[A, Option[A], Value] with SingleModelAccess[A]
+trait SingleRowModelAccess[+A] extends SingleModelAccess[A] with RowFactoryView[A]