+++ Vault/src/utopia/vault/nosql/access/template/model/NonDeprecatedAccess.scala
@@ -1,6 +1,5 @@
-package utopia.vault.nosql.access.template
+package utopia.vault.nosql.access.template.model
 
-import utopia.vault.nosql.access.template.model.ModelAccess
 import utopia.vault.nosql.factory.FromResultFactory
 import utopia.vault.nosql.template.Deprecatable
 
@@ -9,6 +8,7 @@ import utopia.vault.nosql.template.Deprecatable
   * @author Mikko Hilpinen
   * @since 30.1.2020, v1.4
   */
+@deprecated("Replaced with NonDeprecatableView", "v1.8")
 trait NonDeprecatedAccess[+M, +A, +V] extends ModelAccess[M, A, V]
 {
 	// ABSTRACT	------------------------