+++ Vault/src/utopia/vault/nosql/view/FilterableView.scala
@@ -1,14 +1,13 @@
-package utopia.vault.nosql.access.template
+package utopia.vault.nosql.view
 
 import utopia.vault.sql.Condition
 
 /**
   * A common trait for access points whose conditions may be enhanced to cover a more precise area
   * @author Mikko Hilpinen
-  * @since 30.1.2020, v1.4
+  * @since 11.7.2021, v1.8
   */
-@deprecated("Replaced with FilterableView", "v1.8")
-trait FilterableAccess[+A, +Repr] extends Access[A]
+trait FilterableView[+Sub] extends View
 {
 	// ABSTRACT	----------------------
 	
@@ -17,5 +16,5 @@ trait FilterableAccess[+A, +Repr] extends Access[A]
 	  * @param additionalCondition An additional search condition applied
 	  * @return A copy of this access point with specified search condition in place
 	  */
-	def filter(additionalCondition: Condition): Repr
+	def filter(additionalCondition: Condition): Sub
 }