+++ Vault/src/utopia/vault/nosql/view/UnconditionalView.scala
@@ -1,12 +1,11 @@
-package utopia.vault.nosql.access.template
+package utopia.vault.nosql.view
 
 /**
   * A common trait for access points that don't apply any conditions
   * @author Mikko Hilpinen
-  * @since 30.1.2020, v0.1
+  * @since 11.7.2021, v1.8
   */
-@deprecated("Replaced with UnconditionalView", "v1.8")
-trait UnconditionalAccess[+A] extends Access[A]
+trait UnconditionalView extends View
 {
 	// Global condition is forced to be None
 	final override def globalCondition: None.type = None